+++ rest/microvisor/v1/model_list_account_config_response.go
@@ -14,13 +14,8 @@
 
 package openapi
 
-// ListAppResponseMeta struct for ListAppResponseMeta
-type ListAppResponseMeta struct {
-	FirstPageUrl    string  `json:"first_page_url,omitempty"`
-	NextPageUrl     *string `json:"next_page_url,omitempty"`
-	Page            int     `json:"page,omitempty"`
-	PageSize        int     `json:"page_size,omitempty"`
-	PreviousPageUrl *string `json:"previous_page_url,omitempty"`
-	Url             string  `json:"url,omitempty"`
-	Key             string  `json:"key,omitempty"`
+// ListAccountConfigResponse struct for ListAccountConfigResponse
+type ListAccountConfigResponse struct {
+	Configs []MicrovisorV1AccountConfig   `json:"configs,omitempty"`
+	Meta    ListAccountConfigResponseMeta `json:"meta,omitempty"`
 }