+++ rest/microvisor/v1/model_list_device_response.go
@@ -16,6 +16,6 @@ package openapi
 
 // ListDeviceResponse struct for ListDeviceResponse
 type ListDeviceResponse struct {
-	Devices []MicrovisorV1Device `json:"devices,omitempty"`
-	Meta    ListAppResponseMeta  `json:"meta,omitempty"`
+	Devices []MicrovisorV1Device          `json:"devices,omitempty"`
+	Meta    ListAccountConfigResponseMeta `json:"meta,omitempty"`
 }