+++ CHANGES.md
@@ -1,5 +1,12 @@
 twilio-go changelog
 ====================
+[2023-02-08] Version 1.3.3
+--------------------------
+**Lookups**
+- Add `disposable_phone_number_risk` package to the lookup response
+- Add `sms_pumping_risk` package to the lookup response
+
+
 [2023-01-25] Version 1.3.2
 --------------------------
 **Library - Docs**